requirements.txt
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
cycler==0.10.0
future==0.18.2
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.0.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.6
numpy==1.18.1
pycparser==2.19
pylint==2.4.4
pyparsing==2.4.6
python-bidi==0.4.2
python-dateutil==2.8.1
python-telegram-bot==12.4.2
selenium==3.141.0
six==1.14.0
soupsieve==2.0
tornado==6.0.3
urllib3==1.25.8
wrapt==1.11.2
requests==2.23.0
lxml==4.5.0

